Backend/requirement.txt
aniso8601==7.0.0
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
dj-database-url==1.0.0
Django==3.2.5
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
Faker==8.12.1
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
Pillow==8.4.0
promise==2.3
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.1
requests==2.28.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
